requirements-dev.txt
# linting
flake8
flake8-isort
flake8-print
flake8-blind-except
pep8-naming
black

#typing
mypy
django-types
types-pytz

# security
bandit

# testing
factory_boy
pytest
pytest-django
pytest-env
pytest-cov

-r ./requirements.lock
